+++ google-maps/android/src/main/java/com/capacitorjs/plugins/googlemaps/CapacitorGoogleMapsPlugin.kt
@@ -376,6 +376,31 @@ class CapacitorGoogleMapsPlugin : Plugin() {
     }
 
     @PluginMethod
+    fun getMapType(call: PluginCall) {
+        try {
+            val id = call.getString("id")
+            id ?: throw InvalidMapIdError()
+
+            val map = maps[id]
+            map ?: throw MapNotFoundError()
+
+            map.getMapType() { type, err ->
+
+                if (err != null) {
+                    throw err
+                }
+                val data = JSObject()
+                data.put("type", type)
+                call.resolve(data)
+            }
+        } catch (e: GoogleMapsError) {
+            handleError(call, e)
+        } catch (e: Exception) {
+            handleError(call, e)
+        }
+    }
+
+    @PluginMethod
     fun setMapType(call: PluginCall) {
         try {
             val id = call.getString("id")
@@ -572,7 +597,7 @@ class CapacitorGoogleMapsPlugin : Plugin() {
 
             CoroutineScope(Dispatchers.Main).launch {
                 val bounds = map.getLatLngBounds()
-                val data = map.getLatLngBoundsJSObject(bounds)
+                val data = getLatLngBoundsJSObject(bounds)
                 call.resolve(data)
             }
         } catch (e: GoogleMapsError) {
@@ -594,6 +619,27 @@ class CapacitorGoogleMapsPlugin : Plugin() {
                 val contains = bounds.contains(point)
                 val data = JSObject()
                 data.put("contains", contains)
+                call.resolve(data)
+            }
+        } catch (e: GoogleMapsError) {
+            handleError(call, e)
+        } catch (e: Exception) {
+            handleError(call, e)
+        }
+    }
+
+    @PluginMethod
+    fun mapBoundsExtend(call: PluginCall) {
+        try {
+            val boundsObject = call.getObject("bounds")
+            val pointObject = call.getObject("point")
+
+            CoroutineScope(Dispatchers.Main).launch {
+                val bounds = createLatLngBounds(boundsObject)
+                val point = createLatLng(pointObject)
+                val newBounds = bounds.including(point)
+                val data = JSObject()
+                data.put("bounds", getLatLngBoundsJSObject(newBounds))
                 call.resolve(data)
             }
         } catch (e: GoogleMapsError) {