+++ google-maps/src/web.ts
@@ -9,7 +9,7 @@ import {
 } from '@googlemaps/markerclusterer';
 
 import type { Marker } from './definitions';
-import { LatLngBounds } from './definitions';
+import { MapType, LatLngBounds } from './definitions';
 import type {
   AccElementsArgs,
   AddMarkerArgs,
@@ -28,6 +28,7 @@ import type {
   OnScrollArgs,
   MapBoundsContainsArgs,
   EnableClusteringArgs,
+  MapBoundsExtendArgs,
 } from './implementation';
 
 export class CapacitorGoogleMapsWeb
@@ -123,8 +124,23 @@ export class CapacitorGoogleMapsWeb
     });
   }
 
+  async getMapType(_args: { id: string }): Promise<{ type: string }> {
+    let type = this.maps[_args.id].map.getMapTypeId();
+    if (type !== undefined) {
+      if (type === 'roadmap') {
+        type = MapType.Normal;
+      }
+      return { type };
+    }
+    throw new Error('Map type is undefined');
+  }
+
   async setMapType(_args: MapTypeArgs): Promise<void> {
-    this.maps[_args.id].map.setMapTypeId(_args.mapType);
+    let mapType = _args.mapType.toLowerCase();
+    if (mapType === MapType.Normal) {
+      mapType = 'roadmap';
+    }
+    this.maps[_args.id].map.setMapTypeId(mapType);
   }
 
   async enableIndoorMaps(_args: IndoorMapArgs): Promise<void> {
@@ -313,6 +329,29 @@ export class CapacitorGoogleMapsWeb
     return { contains: bounds.contains(point) };
   }
 
+  async mapBoundsExtend(
+    _args: MapBoundsExtendArgs,
+  ): Promise<{ bounds: LatLngBounds }> {
+    const bounds = this.getLatLngBounds(_args.bounds);
+    const point = new google.maps.LatLng(_args.point.lat, _args.point.lng);
+    bounds.extend(point);
+    const result = new LatLngBounds({
+      southwest: {
+        lat: bounds.getSouthWest().lat(),
+        lng: bounds.getSouthWest().lng(),
+      },
+      center: {
+        lat: bounds.getCenter().lat(),
+        lng: bounds.getCenter().lng(),
+      },
+      northeast: {
+        lat: bounds.getNorthEast().lat(),
+        lng: bounds.getNorthEast().lng(),
+      },
+    });
+    return { bounds: result };
+  }
+
   private getLatLngBounds(_args: LatLngBounds): google.maps.LatLngBounds {
     return new google.maps.LatLngBounds(
       new google.maps.LatLng(_args.southwest.lat, _args.southwest.lng),
@@ -449,6 +488,7 @@ export class CapacitorGoogleMapsWeb
       title: marker.title,
       icon: iconImage,
       draggable: marker.draggable,
+      zIndex: marker.zIndex,
     };
 
     return opts;