+++ push-notifications/android/src/main/java/com/capacitorjs/plugins/pushnotifications/PushNotificationsPlugin.java
@@ -1,5 +1,6 @@
 package com.capacitorjs.plugins.pushnotifications;
 
+import android.Manifest;
 import android.app.Notification;
 import android.app.NotificationManager;
 import android.content.Context;
@@ -14,16 +15,20 @@ import androidx.core.app.NotificationCompat;
 import com.getcapacitor.*;
 import com.getcapacitor.annotation.CapacitorPlugin;
 import com.getcapacitor.annotation.Permission;
+import com.getcapacitor.annotation.PermissionCallback;
 import com.google.firebase.messaging.FirebaseMessaging;
 import com.google.firebase.messaging.RemoteMessage;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.List;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-@CapacitorPlugin(name = "PushNotifications", permissions = @Permission(strings = {}, alias = "receive"))
+@CapacitorPlugin(
+    name = "PushNotifications",
+    permissions = @Permission(strings = { Manifest.permission.POST_NOTIFICATIONS }, alias = PushNotificationsPlugin.PUSH_NOTIFICATIONS)
+)
 public class PushNotificationsPlugin extends Plugin {
+
+    static final String PUSH_NOTIFICATIONS = "receive";
 
     public static Bridge staticBridge = null;
     public static RemoteMessage lastMessage = null;
@@ -56,10 +61,9 @@ public class PushNotificationsPlugin extends Plugin {
             JSObject dataObject = new JSObject();
             for (String key : bundle.keySet()) {
                 if (key.equals("google.message_id")) {
-                    notificationJson.put("id", bundle.get(key));
+                    notificationJson.put("id", bundle.getString(key));
                 } else {
-                    Object value = bundle.get(key);
-                    String valueStr = (value != null) ? value.toString() : null;
+                    String valueStr = bundle.getString(key);
                     dataObject.put(key, valueStr);
                 }
             }
@@ -68,6 +72,30 @@ public class PushNotificationsPlugin extends Plugin {
             actionJson.put("actionId", "tap");
             actionJson.put("notification", notificationJson);
             notifyListeners("pushNotificationActionPerformed", actionJson, true);
+        }
+    }
+
+    @PluginMethod
+    public void checkPermissions(PluginCall call) {
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.TIRAMISU) {
+            JSObject permissionsResultJSON = new JSObject();
+            permissionsResultJSON.put("receive", "granted");
+            call.resolve(permissionsResultJSON);
+        } else {
+            super.checkPermissions(call);
+        }
+    }
+
+    @PluginMethod
+    public void requestPermissions(PluginCall call) {
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.TIRAMISU) {
+            JSObject permissionsResultJSON = new JSObject();
+            permissionsResultJSON.put("receive", "granted");
+            call.resolve(permissionsResultJSON);
+        } else {
+            if (getPermissionState(PUSH_NOTIFICATIONS) != PermissionState.GRANTED) {
+                requestPermissionForAlias(PUSH_NOTIFICATIONS, call, "permissionsCallback");
+            }
         }
     }
 
@@ -86,6 +114,13 @@ public class PushNotificationsPlugin extends Plugin {
                     sendToken(task.getResult());
                 }
             );
+        call.resolve();
+    }
+
+    @PluginMethod
+    public void unregister(PluginCall call) {
+        FirebaseMessaging.getInstance().setAutoInitEnabled(false);
+        FirebaseMessaging.getInstance().deleteToken();
         call.resolve();
     }
 
@@ -111,7 +146,7 @@ public class PushNotificationsPlugin extends Plugin {
                     JSObject extras = new JSObject();
 
                     for (String key : notification.extras.keySet()) {
-                        extras.put(key, notification.extras.get(key));
+                        extras.put(key, notification.extras.getString(key));
                     }
 
                     jsNotif.put("data", extras);
@@ -221,14 +256,22 @@ public class PushNotificationsPlugin extends Plugin {
             if (presentation != null) {
                 if (Arrays.asList(presentation).contains("alert")) {
                     Bundle bundle = null;
-                    try {
-                        ApplicationInfo applicationInfo = getContext()
-                            .getPackageManager()
-                            .getApplicationInfo(getContext().getPackageName(), PackageManager.GET_META_DATA);
-                        bundle = applicationInfo.metaData;
-                    } catch (PackageManager.NameNotFoundException e) {
-                        e.printStackTrace();
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+                        try {
+                            ApplicationInfo applicationInfo = getContext()
+                                .getPackageManager()
+                                .getApplicationInfo(
+                                    getContext().getPackageName(),
+                                    PackageManager.ApplicationInfoFlags.of(PackageManager.GET_META_DATA)
+                                );
+                            bundle = applicationInfo.metaData;
+                        } catch (PackageManager.NameNotFoundException e) {
+                            e.printStackTrace();
+                        }
+                    } else {
+                        bundle = getBundleLegacy();
                     }
+
                     int pushIcon = android.R.drawable.ic_dialog_info;
 
                     if (bundle != null && bundle.getInt("com.google.firebase.messaging.default_notification_icon") != 0) {
@@ -267,5 +310,23 @@ public class PushNotificationsPlugin extends Plugin {
             return (PushNotificationsPlugin) handle.getInstance();
         }
         return null;
+    }
+
+    @PermissionCallback
+    private void permissionsCallback(PluginCall call) {
+        this.checkPermissions(call);
+    }
+
+    @SuppressWarnings("deprecation")
+    private Bundle getBundleLegacy() {
+        try {
+            ApplicationInfo applicationInfo = getContext()
+                .getPackageManager()
+                .getApplicationInfo(getContext().getPackageName(), PackageManager.GET_META_DATA);
+            return applicationInfo.metaData;
+        } catch (PackageManager.NameNotFoundException e) {
+            e.printStackTrace();
+            return null;
+        }
     }
 }