+++ screen-orientation/ios/Plugin/ScreenOrientationPlugin.m
@@ -1,0 +1,9 @@
+#import <Foundation/Foundation.h>
+#import <Capacitor/Capacitor.h>
+
+CAP_PLUGIN(ScreenOrientationPlugin, "ScreenOrientation",
+  CAP_PLUGIN_METHOD(orientation, CAPPluginReturnPromise);
+  CAP_PLUGIN_METHOD(lock, CAPPluginReturnPromise);
+  CAP_PLUGIN_METHOD(unlock, CAPPluginReturnPromise);
+  CAP_PLUGIN_METHOD(removeAllListeners, CAPPluginReturnPromise);
+)